==> notebooks/STEP_1_10-K_Extraction.ipynb <==
#installing sec edgar using pip - we will collect our 10-Ks from here
pip install sec-edgar-downloader

# installing necessary libraries, including downloader from sec edgar, allowing us to store our 10-Ks on our local computer.
#using beautiful soup to parse HTML

import os
import time
import pandas as pd
import requests
from bs4 import BeautifulSoup
from sec_edgar_downloader import Downloader
from tqdm import tqdm

# Creating a function that takes in a Wikipedia URL and extracts the stock tickers listed in the table.
# S&P pages use a class="wikitable", so we find the table and then convert it to a DataFrame.
# We also replace '.' with '-' because EDGAR uses dash format.

def fetch_sp_tickers(wiki_url):
    resp = requests.get(wiki_url)
    soup = BeautifulSoup(resp.text, "html.parser")
    table = soup.find('table', {'class': 'wikitable'})
    df = pd.read_html(str(table))[0]
    return df['Symbol'].str.replace('.', '-', regex=False).tolist()

# obtaining the list of 1500 S&P companies from wikipedia
wiki_urls = [
    "https://en.wikipedia.org/wiki/List_of_S&P_500_companies",
    "https://en.wikipedia.org/wiki/List_of_S&P_400_companies",
    "https://en.wikipedia.org/wiki/List_of_S&P_600_companies"
]

#initializing empty list to store the tickers
tickers = []

#running a loop to gather all our tickers
for url in wiki_urls:
    tickers += fetch_sp_tickers(url)
tickers = sorted(set(tickers))
print(f"Total unique tickers: {len(tickers)}")  #expecting around 1500 based on our URLs

#Downloading the 10-Ks by feeding our tickers
#using our email since sec-edgar requires one

dl = Downloader("sec-edgar-filings", "shahk83@mcmaster.ca")
for ticker in tqdm(tickers):
    out_dir = os.path.join("sec-edgar-filings", ticker, "10-K")
    if os.path.exists(out_dir) and os.listdir(out_dir):
        continue  # skip if already downloaded
    try:
        dl.get("10-K", ticker)
    except Exception as e:
        print(f"❌ {ticker} error: {e}")
    time.sleep(0.5)  # using this since there is a rate limit for sec edgar (10/s)


# our output shows that we gathered a total of 1505 tickers from sec-edgar using our list
# some of tickers gave us an error and could not be obtained.
#running this code took us 5 hours
