==> notebooks/STEP_4_convert_to_parquet.ipynb <==
import pandas as pd

df = pd.read_csv("10K_extracted_sections.csv")
df.to_parquet("10K_extracted_sections.parquet", index=False)
print("✅ Converted to Parquet.")


#this file is pretty straight forward, we are converting our results (csv file) to parquet
#parquet compresses the file without losing data
#since our output was very high, we now have smaller file size, improving loading and processing time.

#drawbacks of using csv - when we used csv, our computers started crashing due to the high processing power for the file
