==> notebooks/STEP_6_DL_data_cleaning.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from bs4 import BeautifulSoup

df = pd.read_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix.parquet')

#need to clean the html tags in the columns item1a and item7 to just keep the natural language
df.head()

df.shape

df_copy = df.copy()

df = df_copy.copy()

df.loc[0, 'item1a']

#using beautiful soup library to scrape text from the item1a and item7
from bs4 import BeautifulSoup
from tqdm import tqdm
import pandas as pd
import os

def clean_html(text):
    if pd.isna(text):
        return ""
    try:
        soup = BeautifulSoup(text, "html5lib")
        return soup.get_text(separator=" ", strip=True)
    except Exception as e:
        return str(text)

# here, batch processing is done to optimize the extraction process
# Parameters
batch_size = 50
save_every = 1  # Save every N batches
start_index = 0
output_file = "cleaned_partial.csv"


df['item1a'] = df['item1a'].astype(str)
df['item7'] = df['item7'].astype(str)

# Clean and save in batches
for i in tqdm(range(start_index, len(df), batch_size), desc="🧹 Cleaning batches"):
    print(f"processing index: {i}")
    end = min(i + batch_size, len(df))
    df.loc[i:end, 'item1a'] = df.loc[i:end, 'item1a'].apply(clean_html)
    df.loc[i:end, 'item7'] = df.loc[i:end, 'item7'].apply(clean_html)

#succesfully extracted tex
df.iloc[0, 2]
