==> notebooks/STEP_7_Feature_eng_text_embeddings.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import MultiLabelBinarizer
import ast

df = pd.read_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_clean_final.parquet')
df.shape

df.head()

# there are a few rows where inappropriate text is found like '' or '.' wil have to remove those
df.tail()

#cleaning ultra short text
def clean_short_or_empty(text):
    if isinstance(text, str):
        stripped = text.strip()
        # Remove if empty, only digits, or too short to be meaningful
        if stripped == "" or stripped.isdigit() or len(stripped) < 5:
            return np.nan
    return text

# Apply to cleaned text columns
df['item1a'] = df['item1a'].apply(clean_short_or_empty)
# df['item7']  = df['item7'].apply(clean_short_or_empty)

df.item1a.isna().sum()

df = df.replace({None: np.nan})

nan_values = df[(df.item1a.isna()) & (df.item7.isna())]

nan_values

#dropping all the rows which have both item1a and item7 as nan values
df.drop(nan_values.index, inplace=True)

#save the cleaned version as parquet
df.to_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_clean_final_cleaned.parquet', index=False)

df.shape

# this shows the class imbalance in the dataset extracted
df.risk_score.value_counts()

## Field Preprocessing

df = pd.read_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_clean_final_cleaned.parquet')

# combininf item1a and item 7 both the natural language heavy columns
df['text'] = df['item1a'] + " " + df['item7']

df_copy = df.copy()

df = df_copy.copy()

# Create label (0 to 4 for risk score 1 to 5)
df['label'] = df['risk_score'].astype(int) - 1

#narrative one-hot encoding
df['narrative'].head(10).tolist()

df["text"] = df["item1a"].fillna("").str.cat(df["item7"].fillna(" "), sep=" ")
from sklearn.preprocessing import OneHotEncoder
import pandas as pd

encoder = OneHotEncoder(sparse_output=False)
narr_encoded = encoder.fit_transform(df[['narrative']])

# Create DataFrame from encoded values
narr_encoded_df = pd.DataFrame(narr_encoded, columns=encoder.get_feature_names_out(['narrative']))

# Add to original dataframe
df = pd.concat([df.reset_index(drop=True), narr_encoded_df.reset_index(drop=True)], axis=1)


#combining both the columns which contain natural language for further processing (item1a and item7)
df["text"] = df["item1a"].fillna("").str.cat(df["item7"].fillna(" "), sep=" ")

df.head()

df.drop(['item1a', 'item7'], axis=1, inplace=True)

#storing the final preprocessed parquet
df.to_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_preprocessed.parquet', index=False)

## Text embedding using FinBERT

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import MultiLabelBinarizer
import ast
import os
from transformers import AutoTokenizer
from tqdm import tqdm
import torch
import gc

df = pd.read_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_preprocessed.parquet')

save_path = '/content/drive/MyDrive/DeepLearningProject/Fibert_Batches'
os.makedirs(save_path, exist_ok=True)

# Prepare tokenizer and directory
# processing in the batch of size 4 initially to optimize RAM utilization, because loading all the data in RAM was not possible at once
# batch wise text embeddings, attention masks and input_ids(ids for which attention masks are generated)

tokenizer = AutoTokenizer.from_pretrained("yiyanghkust/finbert-pretrain")
texts = df["text"].fillna("").tolist()
batch_size = 4

# Tokenize in batches and save to disk
for i in tqdm(range(0, len(texts), batch_size)):
    batch_texts = texts[i:i + batch_size]
    encoded = tokenizer(
        batch_texts,
        truncation=True,
        padding="max_length",
        max_length=512,
        return_tensors="pt"
    )

    # Save input_ids and attention_mask separately
    torch.save(encoded["input_ids"], f"{save_path}/input_ids_batch_{i}.pt")
    torch.save(encoded["attention_mask"], f"{save_path}/attention_mask_batch_{i}.pt")


    # Clear memory immediately
    del encoded
    gc.collect()
    torch.cuda.empty_cache()


# since processing with batch size of 4 was not possible due to less RAM, whereas the requirement was more to load that much data
# batch size was reduced to 1 for further samples to occupy less RAM and save separate embeddings for every row
# Prepare tokenizer and directory
tokenizer = AutoTokenizer.from_pretrained("yiyanghkust/finbert-pretrain")
texts = df["text"].fillna("").tolist()
batch_size = 4
process_batch = 1
start_batch = 48
start_index = start_batch * batch_size

for i in tqdm(range(start_index, len(texts), process_batch)):
    batch_texts = texts[i:i + process_batch]
    encoded = tokenizer(
        batch_texts,
        truncation=True,
        padding="max_length",
        max_length=512,
        return_tensors="pt"
    )

    # Save input_ids and attention_mask separately
    torch.save(encoded["input_ids"], f"{save_path}/input_ids_batch_{i}.pt")
    torch.save(encoded["attention_mask"], f"{save_path}/attention_mask_batch_{i}.pt")

    # Clear memory
    del encoded
    gc.collect()
    torch.cuda.empty_cache()
