==> notebooks/STEP_8_Parquets_division_batchWise.ipynb <==
This file is dividing the final parquet: risk_matrix_preprocessed.parquet similar to the batch size of which embedings is saved i.e.

batch size 4: for 0-184

batch size 1: for 192-till the end

from google.colab import drive
drive.mount('/content/drive')

path = "/content/drive/MyDrive/DeepLearningProject/Parquet_Batches"

df = pd.read_parquet('/content/drive/MyDrive/DeepLearningProject/risk_matrix_preprocessed.parquet')

df['risk_score'] = df['risk_score'] - 1

df.head()

df.value_counts('risk_score')

import pandas as pd
import os

batch_size_map = {}

# First batch range (batch size = 4)
for i in range(0, 47 * 4, 4):
    batch_size_map[i] = 4

# Second batch range (batch size = 1)
for i in range(192, 515):
    batch_size_map[i] = 1

# os.makedirs("parquet_batches", exist_ok=True)
#saves the new batch wise parquet with showing the starting row of the batch in name of the file
for batch_start, bsize in batch_size_map.items():
    batch_df = df.iloc[batch_start:batch_start + bsize]
    batch_df.to_parquet(f"{path}/risk_matrix_{batch_start}.parquet", index=False)
